app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.5",
    "telethon>=1.40.0",
    "minio>=7.2.15",
    "elasticsearch>=9.0.1",
    "aiohttp>=3.12.0",
    "async-to-sync>=0.2.3",
    "asgiref>=3.8.1",
    "pytest>=8.3.5",
    "httpx>=0.28.1",
    "pytest-asyncio>=1.0.0",
    "fakeredis>=2.29.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "celery>=5.5.2",
    "redis>=6.1.0",
    "flower>=2.0.1",
    "prometheus-client>=0.22.0",
    "setuptools>=80.9.0",
]
